+++ src/core_modules/capture-core/components/Buttons/index.js
@@ -1,2 +1,3 @@
 // @flow
 export { SimpleSplitButton } from './SimpleSplitButton.component';
+export { OverflowButton } from './OverflowButton.component';
